+++ src/speech_to_speech/generic/__init__.py
@@ -1,0 +1,1 @@
+"""Generic Speech-to-Speech example package."""